Introduce a config parameter: urlregex
Allows for a regex to be applied to each article url when first encountered in a feed.
Born from noticing visir.is editing errors in headlines by changing the url.

Example:

http://www.visir.is/g/2017170629354/glaepagengi-i-arhusum-ogna-oryggi-almennings

Changed to:

http://www.visir.is/g/2017170629354/glaepagengi-i-arosum-ogna-oryggi-almennings

With a regex that only captures ‘http://www.visir.is/g/2017170629354’ (which is enough for the article to be displayed) then such a change is picked up.

Example config:

urlregex: (.*\/g\/\d*)