Config kill_classes
Supply a list of classes that are removed with lxml prior to readability doing it’s parsing.
It’s messy but will suffice for now.
Prompted because readability does not properly parse pages like this:

http://bleikt.pressan.is/lesa/dottir-kallar-mommu-sina-feita-svar-modur-hennar-hefur-vakid-mikla-athygli/

(It Is picking up excerpts at the bottom - with the config: ‘kill_classes: ["articles-with-excerpt”]’ then we are fine)